Customized tabs, moved state down
