+++ src/js/ui/NewDataExploration/AdvancedConstraintCheckbox.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
-import {isComponentRerendering} from "../TabSystem";
+import {isComponentRerendering} from "./Workspace";
 
 function updateLayerConstraints(activeConstraints, index) {
     let tempActiveConstraints = [...activeConstraints];
